requirements.txt
pytesseract
Pillow
requests
watchdog
